fix: address PRRT_kwDOSJAM6s59pVz1 - move capacity tests
