+++ tests/unit/api/test_metrics.py
@@ -2,20 +2,16 @@
 
 from __future__ import annotations
 
-from collections.abc import AsyncIterator
 from datetime import UTC, datetime, timedelta
-from typing import Any
 
 import pytest
-from httpx import ASGITransport, AsyncClient
+from httpx import AsyncClient
 from sqlalchemy.ext.asyncio import AsyncEngine
 
-from awf.api.app import configure_database, create_app
-from awf.common.config import Settings, get_settings
+from awf.common.config import get_settings
 from awf.db.enums import FailureReason, WorkspaceStatus
 from awf.db.repositories import WorkspaceRepository
 from awf.db.session import make_session_factory
-from awf.service.disk import DiskCheck
 
 
 async def _workspace(
@@ -43,40 +39,6 @@ async def _workspace(
 
 def _zero_status_counts() -> dict[str, int]:
     return {status.value: 0 for status in WorkspaceStatus}
-
-
-def _disk_check(
-    settings: Settings,
-    *,
-    ok: bool = True,
-    free_bytes: int | None = None,
-    reason: str = "SUFFICIENT_DISK",
-) -> DiskCheck:
-    threshold = settings.min_free_disk_bytes
-    free = threshold + 1 if free_bytes is None else free_bytes
-    return DiskCheck(
-        path=settings.work_dir,
-        checked_path=settings.work_dir,
-        total_bytes=max(free + 1, threshold + 1),
-        used_bytes=1,
-        free_bytes=free,
-        percent_free=99.0,
-        threshold_bytes=threshold,
-        ok=ok,
-        status="ok" if ok else "fail",
-        reason=reason,
-        detail=None if ok else "Free disk is below the configured admission threshold.",
-    )
-
-
-@pytest.fixture
-async def metrics_app_and_client(
-    engine: AsyncEngine,
-) -> AsyncIterator[tuple[Any, AsyncClient]]:
-    app = create_app(use_lifespan=False)
-    configure_database(app, make_session_factory(engine))
-    async with AsyncClient(transport=ASGITransport(app=app), base_url="http://test") as c:
-        yield app, c
 
 
 @pytest.mark.unit
@@ -222,102 +184,3 @@ async def test_workspace_summary_is_token_free_when_api_token_is_configured(
         get_settings.cache_clear()
 
     assert response.status_code == 200
-
-
-@pytest.mark.unit
-async def test_resource_saturation_endpoint_reports_local_capacity_inputs(
-    metrics_app_and_client: tuple[Any, AsyncClient],
-    engine: AsyncEngine,
-) -> None:
-    app, client = metrics_app_and_client
-    settings = Settings(
-        _env_file=None,
-        work_dir="/tmp/awf-metrics-work",
-        min_free_disk_bytes=700,
-        worker_max_concurrent_provisions=5,
-        worker_max_concurrent_executions=2,
-        workspace_steady_cpu=2.0,
-        workspace_steady_memory_gb=7.5,
-        workspace_peak_cpu=4.0,
-        workspace_peak_memory_gb=12.0,
-    )
-    app.dependency_overrides[get_settings] = lambda: settings
-    app.state.workspace_admission_disk_check = lambda provider_settings: _disk_check(
-        provider_settings,
-        ok=True,
-    )
-    now = datetime.now(UTC)
-    for status in (
-        WorkspaceStatus.ready,
-        WorkspaceStatus.running,
-        WorkspaceStatus.validating,
-        WorkspaceStatus.completed,
-    ):
-        await _workspace(engine, status=status, updated_at=now - timedelta(minutes=5))
-
-    response = await client.get("/v1/metrics/resources/saturation")
-
-    assert response.status_code == 200
-    body = response.json()
-    assert body["workspace_counts"]["active_total"] == 3
-    assert body["workspace_counts"]["ready"] == 1
-    assert body["workspace_counts"]["running"] == 1
-    assert body["workspace_counts"]["validating"] == 1
-    assert body["workspace_counts"]["by_status"][WorkspaceStatus.completed.value] == 1
-    assert body["worker"] == {
-        "max_concurrent_provisions": 5,
-        "max_concurrent_executions": 2,
-    }
-    assert body["resource_defaults"] == {
-        "steady_cpu": 2.0,
-        "steady_memory_gb": 7.5,
-        "peak_cpu": 4.0,
-        "peak_memory_gb": 12.0,
-    }
-    assert body["reserved_resources"] == {
-        "active_workspace_count": 3,
-        "steady_cpu": 6.0,
-        "steady_memory_gb": 22.5,
-        "peak_cpu": 12.0,
-        "peak_memory_gb": 36.0,
-    }
-    assert body["concurrency"]["execution"] == {
-        "limit": 2,
-        "in_use": 2,
-        "queued": 1,
-        "available": 0,
-    }
-    assert body["disk"]["reason"] == "SUFFICIENT_DISK"
-    assert body["admission"]["ok"] is True
-    assert body["admission"]["status"] == "saturated"
-    assert body["admission"]["reason"] == "WORKER_EXECUTION_CONCURRENCY_SATURATED"
-
-
-@pytest.mark.unit
-async def test_resource_saturation_endpoint_explains_disk_admission_denial(
-    metrics_app_and_client: tuple[Any, AsyncClient],
-) -> None:
-    app, client = metrics_app_and_client
-    settings = Settings(
-        _env_file=None,
-        work_dir="/tmp/awf-metrics-work",
-        min_free_disk_bytes=700,
-    )
-    app.dependency_overrides[get_settings] = lambda: settings
-    app.state.workspace_admission_disk_check = lambda provider_settings: _disk_check(
-        provider_settings,
-        ok=False,
-        free_bytes=100,
-        reason="INSUFFICIENT_DISK",
-    )
-
-    response = await client.get("/v1/metrics/resources/saturation")
-
-    assert response.status_code == 200
-    body = response.json()
-    assert body["disk"]["ok"] is False
-    assert body["disk"]["reason"] == "INSUFFICIENT_DISK"
-    assert body["admission"]["ok"] is False
-    assert body["admission"]["status"] == "blocked"
-    assert body["admission"]["reason"] == "INSUFFICIENT_DISK"
-    assert "free disk" in body["admission"]["detail"].lower()
